Adding possibility to run locally

+++ local-runner.sh
@@ -2,6 +2,13 @@ SCRIPT_PATH=$(dirname "$0")
 
 cd ${SCRIPT_PATH}
 
+if [ -z $1 ]; then
+    echo "Provide a path to project-{name}.yaml file"
+    exit 1
+fi
+
+export PRJECT_PATH=$1
+
 docker rm -f $(docker-compose ps -a -q)
 sudo rm -rf .data/
 sudo rm -rf dist/
